requirements.txt
anyio==3.6.2
autopep8==2.0.0
certifi==2022.9.24
flake8==5.0.4
fuzzywuzzy==0.18.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.4
importlib-metadata==4.2.0
Levenshtein==0.20.8
mccabe==0.7.0
pyaes==1.6.1
pyasn1==0.4.8
pycodestyle==2.9.1
pyflakes==2.5.0
python-dotenv==0.21.0
python-Levenshtein==0.20.8
rapidfuzz==2.13.0
rfc3986==1.5.0
rsa==4.9
sniffio==1.3.0
Telethon==1.25.4
tomli==2.0.1
typing_extensions==4.4.0
zipp==3.10.0
